add components for test
