+++ .stylelintrc.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  rules: {
+    'at-rule-no-unknown': null,
+  },
+}
