+++ src/app/[locale]/layout.tsx
@@ -4,6 +4,7 @@ import React from 'react'
 import { Geist, Geist_Mono } from 'next/font/google'
 import type { Metadata } from 'next'
 import '../globals.css'
+import 'car-robots-library/dist/style.css'
 import { routing } from '@/src/i18n/routing'
 
 const geistSans = Geist({
