+++ src/shared/components/localeSwither/LocaleSwitcher.tsx
@@ -1,25 +1,48 @@
-import { useLocale, useTranslations } from 'next-intl'
-import React from 'react'
-import LocaleSwitcherSelect from './LocaleSwitcherSelect'
+'use client'
+
+import { Locale, useLocale, useTranslations } from 'next-intl'
+import React, { useTransition } from 'react'
 import { routing } from '@/src/i18n/routing'
+import { Select, SelectItem } from 'car-robots-library'
+import { usePathname, useRouter } from '@/src/i18n/navigation'
+import { useParams } from 'next/navigation'
 
 export default function LocaleSwitcher() {
   const t = useTranslations('HomePage')
   const locale = useLocale()
 
+  const router = useRouter()
+  const [isPending, startTransition] = useTransition()
+  const pathname = usePathname()
+  const params = useParams()
+
+  function onSelectChange(value: string) {
+    const nextLocale = value as Locale
+    startTransition(() => {
+      router.replace(
+        // @ts-expect-error -- TypeScript will validate that only known `params`
+        // are used in combination with a given `pathname`. Since the two will
+        // always match for the current route, we can skip runtime checks.
+        { pathname, params },
+        { locale: nextLocale }
+      )
+    })
+  }
+
   return (
-    <LocaleSwitcherSelect
+    <Select
       defaultValue={locale}
-      label={t('label')}
+      disabled={isPending}
+      onValueChange={onSelectChange}
     >
       {routing.locales.map((current) => (
-        <option
+        <SelectItem
           key={current}
           value={current}
         >
           {t('locale', { locale: current ?? 'ru' })}
-        </option>
+        </SelectItem>
       ))}
-    </LocaleSwitcherSelect>
+    </Select>
   )
 }
